x86: do page table switching when entering/leaving hypervisor
For XPTI enabled domains do page table switching when entering or
leaving the hypervisor. This requires both %cr3 values to be stored
in the per-vcpu stack regions and adding the switching code to the
macros used to switch stacks.

The hypervisor will run on the original L4 page table supplied by the
guest, while the guest will use the shadow.

Signed-off-by: Juergen Gross <jgross@suse.com>