xen: sched/credit1: make vcpu_migration_delay per-cpupool
Right now, vCPU migration delay is controlled by
the vcpu_migration_delay boot parameter. This means:
- it can only be set at boot time, and can't be
  changed;
- it applies to every possible instance of Credit1,
  in any cpupool that will be created.

Add a vcpu_migr_delay field inside struct csched_private,
so that the migration delay:
- we can get/set it online, using XEN_SYSCTL_SCHEDOP_*;
- we can get/set it indepently for each Credit1
  instance, in different cpupools.

The value of the boot time parameter is used for
initializing the vcpu_migr_delay field of the private
structure of all the scheduler instances, when they're
created.

While there, save reading NOW() and doing any s_time_t
operation, when the migration delay of a scheduler is
zero (as it is, by default), in
__csched_vcpu_is_cache_hot().

Finally, note that, from this commit on, using `xenpm
{set,get}-vcpu-migration-delay' is not effective any
longer.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
---
Cc: George Dunlap <george.dunlap@eu.citrix.com>
Cc: Andrew Cooper <andrew.cooper3@citrix.com>