xen: Credit1: do not always tickle an idle pcpu
If a vcpu wakes up on an idle pcpu, we always (try to) run it there.
This may seem ok but it may actually be not, in case there is another
(also idle) pcpu which would be better suited (for load balancing and/or
affinity reasons, for instance) to run it.

So, instead than blindly tickle the pcpu where the vcpu wakes up, let's
go through the tickling, even if it was idle. If there is no 'better'
pcpu where to run it, we still prefer the one where it woke up, among
others with similar characteristics.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
---
Cc: George Dunlap <george.dunlap@eu.citrix.com>