xen: credit2: fix credit reset happening too few times
There is a bug in commit 5e4b4199667b9 ("xen: credit2: only reset
credit on reset condition"). In fact, the aim of that commit was to
make sure that we do not perform too many credit reset operations
(which are not super cheap, and in an hot-path). But the check used
to determine whether a reset is necessary was the wrong one.

In fact, knowing just that some vCPUs have been skipped, while
traversing the runqueue (in runq_candidate()), is not enough. We
need to check explicitly whether the first vCPU in the runqueue
has a negative amount of credit.

Since a trace record is changed, this patch updates xentrace format file
and xenalyze as well

This should be backported.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
---
Cc: George Dunlap <george.dunlap@citrix.com>
Cc: Juergen Gross <jgross@suse.com>
Cc: Jan Beulich <jbeulich@suse.com>
Cc: Charles Arnold <carnold@suse.com>
Cc: Glen <glenbarney@gmail.com>
Cc: Tomas Mozes <hydrapolic@gmail.com>
Cc: Sarah Newman <srn@prgmr.com>
---
About the Credit2 stall issue reported recently, and mentioned in patch
1 of this series. This second patch, alone, was already mitigating the
issue quite substantially.

Still, the proper fix for the issue itself is patch 1, while this is a
fix for a bug in the code, introduced with a previous change, which
happens to help to cure the sympthoms of the problem at hand.